Tidy PR
 - use cArrayElementType in one more place
 - allow allocation of e.g. a 5GB c_array type through c_malloc
   (so cArrayLength returns an int64)
 - remove commented-out flag

+++ compiler/codegen/type.cpp
@@ -248,7 +248,7 @@ void AggregateType::codegenDef() {
       fprintf(outfile, "typedef ");
       fprintf(outfile, "%s", eltTypeCName);
       fprintf(outfile, " %s", symbol->codegen().c.c_str());
-      fprintf(outfile, "[%d];\n\n", (int) sizeInt);
+      fprintf(outfile, "[%ld];\n\n", (long int) sizeInt);
       return;
     } else {
 #ifdef HAVE_LLVM
